+++ package.json
@@ -21,6 +21,7 @@
     "@types/node": "20.4.0",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
+
     "autoprefixer": "10.4.14",
     "class-variance-authority": "^0.6.1",
     "clsx": "^1.2.1",
@@ -28,7 +29,7 @@
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.9",
     "file-saver": "^2.0.5",
-    "framer-motion": "^10.16.1",
+
     "immutability-helper": "^3.1.1",
     "install": "^0.13.0",
     "lucide-react": "^0.259.0",
