+++ src/app/globals.css
@@ -2,7 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
- 
 @layer base {
   :root {
     --background: 0 0% 100%;
