--- src/hooks/useMediaQuery.ts
@@ -1,37 +1,0 @@
-import { useEffect, useState } from "react";
-
-export function useMediaQuery(query: any) {
-  const [matches, setMatches] = useState<boolean>(false);
-
-  useEffect(() => {
-    const media = window.matchMedia(query);
-    if (media.matches !== matches) {
-      setMatches(media.matches);
-    }
-
-    const listener = () => {
-      setMatches(media.matches);
-    };
-
-    if (typeof media.addEventListener === "function") {
-      media.addEventListener("change", listener);
-    } else {
-      media.addListener(listener);
-    }
-
-    return () => {
-      if (typeof media.removeEventListener === "function") {
-        media.removeEventListener("change", listener);
-      }
-      //  else {
-      //     media.removeListener(listenerList);
-      //   }
-    };
-  }, [matches, query]);
-
-  return matches;
-}
-
-export const useIsSmall = () => useMediaQuery("(min-width: 480px)");
-export const useIsMedium = () => useMediaQuery("(min-width: 768px)");
-export const useIsLarge = () => useMediaQuery("(min-width: 1024px)");
